form validation y arreglos estéticos

+++ js/registrocalculo.js
@@ -1,7 +1,7 @@
 export default class RegistroCalculo {
     constructor(monto = 0, tasa = 0, cuotas = 1) {
         this.monto = parseInt(monto);
-        this.tasa = parseInt(tasa);
+        this.tasa = parseFloat(tasa);
         this.cuotas = parseInt(cuotas);
         this.pagoMensual = 0;
         this.pagos = [];
